refactor(scores): split v3 schemas into folder matching v1/v2 layout
Replaces the flat packages/shared/src/features/scores/interfaces/api/scores-v3.ts
with a v3/ folder containing schemas.ts and endpoints.ts, matching the v1/ and
v2/ layout already in this directory. Renames GetScoresV3 to GetScoresQueryV3
for naming parity with GetScoresQueryV1 and GetScoresQueryV2. The private
foundation schema becomes ScoreFoundationSchemaV3 (was ScoreBaseV3) for the same
reason.

No behavior change; consumers using the @langfuse/shared barrel continue to
resolve the public symbols (APIScoreV3, APIScoreSchemaV3, GetScoresResponseV3)
unchanged.
