+++ packages/shared/src/features/scores/interfaces/api/index.ts
@@ -3,4 +3,5 @@ export * from "./v1/validation";
 export * from "./v1/endpoints";
 export * from "./v2/validation";
 export * from "./v2/endpoints";
-export * from "./scores-v3";
+export * from "./v3/schemas";
+export * from "./v3/endpoints";
